+++ game/levels/lvl1.tmx
@@ -13,7 +13,7 @@
  </layer>
  <layer id="9" name="walls" width="20" height="12">
   <data encoding="base64">
-   AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAC4AAKAyAADAMgAAwDIAAMAuAADAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAuAACgMgAAwDIAAMAyAADAMgAAwDIAAMAuAADAMgAAoAAAAAAAAAAAAAAAADMAAMAyAADAMgAAwDIAAMAyAADAMgAAwDIAAMAyAADAMgAAwC8AAKAAAAAAAAAAAAAAAAAAAAAAAAAAADIAAGAyAACgAAAAAAAAAAAAAAAAMQAAoAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA0AABgAAAAAAAAAAAAAAAAMgAAYDIAAKAAAAAAAAAAAAAAAAAxAACgAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAADEAAKAAAAAAAAAAAAAAAAAyAABgMgAAoAAAAAAAAAAAAAAAADEAAKAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAMQAAoAAAAAAAAAAAAAAAADIAAGAyAACgAAAAAAAAAAAAAAAANAAAoAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAxAACgAAAAAAAAAAAAAAAAMgAAYDIAAKAAAAAAAAAAAAAAAAAAAAAAAAAAAC8AAGAyAAAAMgAAADIAAAAyAAAAMgAAADIAAAAyAAAAMgAAADMAAAAAAAAAAAAAAAAAAAAyAABgLgAAADIAAAAyAAAAMgAAADIAAAAyAAAALgAAYAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAALgAAADIAAAAyAAAAMgAAAC4AAGAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA
+   AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAC4AAKAyAADAMgAAwDIAAMAuAADAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAuAACgMgAAwDIAAMAyAADAMgAAwDIAAMAuAADAMgAAoAAAAAAAAAAAAAAAADMAAMAyAADAMgAAwDIAAMAyAADAMgAAwDIAAMAyAADAMgAAwC8AAKAAAAAAAAAAAAAAAAAAAAAAAAAAADIAAGAyAACgAAAAAAAAAAAAAAAAMQAAoAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAwAAAAAAAAAAAAAAAAAAAAMgAAYDIAAKAAAAAAAAAAAAAAAAAxAACgAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAADEAAKAAAAAAAAAAAAAAAAAyAABgMgAAoAAAAAAAAAAAAAAAADEAAKAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAMQAAoAAAAAAAAAAAAAAAADIAAGAyAACgAAAAAAAAAAAAAAAAMAAAwAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAxAACgAAAAAAAAAAAAAAAAMgAAYDIAAKAAAAAAAAAAAAAAAAAAAAAAAAAAAC8AAGAyAAAAMgAAADIAAAAyAAAAMgAAADIAAAAyAAAAMgAAADMAAAAAAAAAAAAAAAAAAAAyAABgLgAAADIAAAAyAAAAMgAAADIAAAAyAAAALgAAYAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAALgAAADIAAAAyAAAAMgAAAC4AAGAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA
   </data>
  </layer>
  <layer id="8" name="checkpoint" width="20" height="12">
