+++ game/levels/lvl5.tmx
@@ -13,7 +13,7 @@
  </layer>
  <layer id="3" name="walls" width="20" height="12">
   <data encoding="base64">
-   LgAAoDIAAMAyAADAMgAAwDIAAMAyAADAMgAAwDIAAMAyAADAMgAAwDIAAMAyAADAMgAAwDIAAMAyAADAMgAAwDIAAMAyAADALgAAwAAAAAAyAACgAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAyAABgAAAAAD0AAMAxAAAAMQAAADEAAAAxAAAAMQAAADEAAAAxAAAAMQAAADEAAAAxAAAAMQAAADEAAAAxAAAAMQAAADwAAAAAAAAALwAAYC4AAGAAAAAAMgAAoAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAMQAAoAAAAAAyAABgAAAAAAAAAAAuAAAAMgAAAC8AAAAAAAAAPAAAYDEAAAAxAAAAMQAAADEAAAAxAAAAMQAAADEAAAAxAAAAPAAAAAAAAAAxAACgAAAAADIAAGAAAAAAAAAAAAAAAAAAAAAAMgAAoAAAAAAxAABgAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAxAACgAAAAADEAAKAAAAAAMgAAYAAAAAAAAAAAAAAAAAAAAAAyAACgAAAAADEAAGAAAAAANAAAYAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAADEAAKAAAAAAMQAAoAAAAAAyAABgAAAAAAAAAAAAAAAAAAAAADIAAKAAAAAAMQAAYAAAAAA8AADAMQAAwDEAAMAxAADAMQAAwDEAAMAxAADAPAAAoAAAAAAxAACgAAAAADIAAGAAAAAAAAAAAAAAAAAAAAAAMgAAoAAAAAAxAABgAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAADEAAKAAAAAAMgAAYAAAAAAAAAAAAAAAAAAAAAAyAACgAAAAADwAAMAxAADAMQAAwDEAAMAxAADAMQAAwDEAAMAxAADAMQAAwDEAAMAxAADAPAAAoAAAAAAyAABgAAAAAAAAAAAAAAAAAAAAADIAAKAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAADIAAGAAAAAAAAAAAAAAAAAAAAAALgAAADIAAAAyAAAAMgAAADIAAAAyAAAAMgAAADIAAAAyAAAAMgAAADIAAAAyAAAAMgAAADIAAAAyAAAALgAAYAAAAAAAAAAA
+   LgAAoDIAAMAyAADAMgAAwDIAAMAyAADAMgAAwDIAAMAyAADAMgAAwDIAAMAyAADAMgAAwDIAAMAyAADAMgAAwDIAAMAyAADALgAAwAAAAAAyAACgAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAyAABgAAAAAD0AAMAxAAAAMQAAADEAAAAxAAAAMQAAADEAAAAxAAAAMQAAADEAAAAxAAAAMQAAADEAAAAxAAAAMQAAADwAAAAAAAAALwAAYC4AAGAAAAAAMgAAoAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAMQAAoAAAAAAyAABgAAAAAAAAAAAuAAAAMgAAAC8AAAAAAAAAPAAAYDEAAAAxAAAAMQAAADEAAAAxAAAAMQAAADEAAAAxAAAAPAAAAAAAAAAxAACgAAAAADIAAGAAAAAAAAAAAAAAAAAAAAAAMgAAoAAAAAAxAABgAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAxAACgAAAAADEAAKAAAAAAMgAAYAAAAAAAAAAAAAAAAAAAAAAyAACgAAAAADEAAGAAAAAAMAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAADEAAKAAAAAAMQAAoAAAAAAyAABgAAAAAAAAAAAAAAAAAAAAADIAAKAAAAAAMQAAYAAAAAA8AADAMQAAwDEAAMAxAADAMQAAwDEAAMAxAADAPAAAoAAAAAAxAACgAAAAADIAAGAAAAAAAAAAAAAAAAAAAAAAMgAAoAAAAAAxAABgAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAADEAAKAAAAAAMgAAYAAAAAAAAAAAAAAAAAAAAAAyAACgAAAAADwAAMAxAADAMQAAwDEAAMAxAADAMQAAwDEAAMAxAADAMQAAwDEAAMAxAADAPAAAoAAAAAAyAABgAAAAAAAAAAAAAAAAAAAAADIAAKAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAADIAAGAAAAAAAAAAAAAAAAAAAAAALgAAADIAAAAyAAAAMgAAADIAAAAyAAAAMgAAADIAAAAyAAAAMgAAADIAAAAyAAAAMgAAADIAAAAyAAAALgAAYAAAAAAAAAAA
   </data>
  </layer>
  <layer id="4" name="checkpoint" width="20" height="12">
