+++ game/levels/mymap.tsx
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<tileset version="1.4" tiledversion="1.4.3" name="mymap" tilewidth="50" tileheight="50" tilecount="19" columns="0">
+<tileset version="1.4" tiledversion="1.4.3" name="mymap" tilewidth="50" tileheight="50" tilecount="18" columns="0">
  <grid orientation="orthogonal" width="1" height="1"/>
  <tile id="44">
   <image width="25" height="25" source="../assets/YELLOW_DOT.png"/>
@@ -21,9 +21,6 @@
  </tile>
  <tile id="50">
   <image width="50" height="50" source="../assets/RwTC.png"/>
- </tile>
- <tile id="51">
-  <image width="50" height="50" source="../assets/RTB.png"/>
  </tile>
  <tile id="52">
   <image width="50" height="50" source="../assets/RLTB.png"/>
